ExactPredicates.jl

ExactPredicates.jl

This package provides fast and robust predicates for Euclidean geometry, including orient, incircle and insphere.

Robustness

Robust means that the code:

raises an exception on NaN and Inf arguments;

gives a correct answer on all other inputs with Float64 coordinates, no matter what (overflow, underflow, etc.);

in particular, no restriction on the coordinate range.

Why robustness matter?

Even if the geometric data is approximate (for example when it comes from measurement), robust computation is important because it guarantees soundness with respect to some combinatorial properties of the predicates. For example orient(a, b, c) == orient(b, c, a) == orient(c, a, b).

“Inexact versions of these tests [orient and incircle] are vulnerable to roundoff error, and the wrong answers they produce can cause geometric algorithms to hang, crash, or produce incorrect output.”

Jonathan Shewchuk, Robust Adaptive Floating-point Geometric Predicates

Type for points

The basic type for representing points is NTuple{N, Float64}, where N is 2 or 3. To define the predicates for a type T, simply define a function Tuple(::T) or coord(::T) that output a NTuple{Float64, Float64} that contains the coordinates. Naturally, the computation is only robust if the conversion is robust too. There should be no performance penalty in the conversion.

using ExactPredicates
struct Point
    x :: Float64
    y :: Float64
end

Tuple(p :: Point) = (p.x, p.y)
incircle(Point(0.0, 0.0), Point(1.0, 0.0), Point(0.0, 1.0), Point(.5, .5))

coord(p :: Float64) = (p, 0.0)
coord(p :: Complex) = reim(p)
incircle(0.0, 1.0, complex(0.0, 1.0), complex(.5, .5))

using StaticArrays
# StaticArrays already defines Tuple for SVector
incircle(SVector(0.0, 0.0), SVector(1.0, 0.0), SVector(0.0, 1.0), SVector(.5, .5))

License

The package is released under the MIT license.

Exported functions

ExactPredicates.closestpoint — Method.

closestpoint(p :: 3, q :: 3, a :: 3) -> Int

Return 1 if a is closer to p than to q. Return –1 if a is closer to q than to p. Return 0 is a is equaly close to both.

source

ExactPredicates.closestpoint — Method.

closestpoint(p :: 2, q :: 2, a :: 2) -> Int

Return 1 if a is closer to p than to q.

Return –1 if a is closer to q than to p.

Return 0 is a is equaly close to both.

source

ExactPredicates.incircle — Method.

incircle(a :: 2, b :: 2, c :: 2, p :: 2) -> Int

Assume that a, b and c define a counterclockwise triangle.

Return 1 if p is strictly inside the circumcircle of this triangle.

Return –1 if p is outside.

Return 0 if p is on the circle.

If the triangle is oriented clockwise, the signs are reversed. If a, b and c are collinear, this degenerate to an orientation test.

If two of the four arguments are equal, return 0.

source

ExactPredicates.insphere — Method.

insphere(p :: 3, q :: 3, r :: 3, s :: 3, a :: 3)

Return 1 if a is inside the circumscribed sphere defined by the four points p, q, r and s.

Return –1 if a is outside.

Return 0 is a lies on the sphere or if the four points are coplanar.

source

ExactPredicates.orient — Method.

orient(p :: 3, q :: 3, r :: 3, a :: 3) -> Int

Consider the oriented plane on which the triangle pqr is positively oriented.

Return 1 if a is below this plane.

Return –1 if a is above this plane.

Return 0 if a lies on this plane.

source

ExactPredicates.orient — Method.

orient(p :: 2, q :: 2, r :: 2) -> Int

Return 1 if r is on the left of the oriented line defined by p and q.

Return –1 if r is on the right.

Return 0 if r is on the line or if p == q.

source